Organizamos el codigo para que sea mas entendible
